+++ src/frontend/src/reducers/neighborhoods/neighborhoodsThunk.ts
@@ -3,6 +3,14 @@ import {
     fetchNeighborhoods,
     fetchNeighborhoodBounds,
 } from "../../api/neighborhoods";
+import {
+    selectNeighborhoodNameByZipcode,
+    setRankedNeighborhoodLists,
+} from "./neighborhoodsSlice";
+import { type AppDispatch, type RootState } from "store/store";
+import neighborhoodsSortedWithRoutes from "./selectors/neighborhoodsSortedWithRoutes";
+import type { RankedNeighborhoodsLists } from "./types";
+import groupRankingsByLikeNeigborhoodName from "./utils/groupRankingListsByName";
 
 export const getNeighborhoodsAndBounds = createAsyncThunk(
     "neighborhoods/getNeighborhoodsAndBounds",
@@ -19,3 +27,44 @@ export const getNeighborhoodsAndBounds = createAsyncThunk(
         return { neighborhoods, neighborhoodBounds };
     }
 );
+
+export const getRankedNeighborhoodLists =
+    () => (dispatch: AppDispatch, getState: () => RootState) => {
+        const state = getState() as RootState;
+
+        const groupedNeighborhoodsLists: RankedNeighborhoodsLists = {
+            topTen: [],
+            groupedTopTen: [],
+            groupedRecommended: [],
+            groupedTooFar: [],
+        };
+
+        const neighborhoodsList = neighborhoodsSortedWithRoutes(state);
+        const neighborhoodNameByZipcode =
+            selectNeighborhoodNameByZipcode(state);
+
+        groupedNeighborhoodsLists.topTen = neighborhoodsList.recommended.slice(
+            0,
+            10
+        );
+
+        if (neighborhoodNameByZipcode) {
+            groupedNeighborhoodsLists.groupedTopTen =
+                groupRankingsByLikeNeigborhoodName(
+                    groupedNeighborhoodsLists.topTen,
+                    neighborhoodNameByZipcode
+                );
+            groupedNeighborhoodsLists.groupedTooFar =
+                groupRankingsByLikeNeigborhoodName(
+                    neighborhoodsList.tooFar,
+                    neighborhoodNameByZipcode
+                );
+            groupedNeighborhoodsLists.groupedRecommended =
+                groupRankingsByLikeNeigborhoodName(
+                    neighborhoodsList.recommended.slice(10),
+                    neighborhoodNameByZipcode
+                );
+        }
+
+        dispatch(setRankedNeighborhoodLists(groupedNeighborhoodsLists));
+    };