+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Migrate route and travel time logic [#693](https://github.com/azavea/echo-locator/pull/693)
 - Add neighborhood ranking logic [#694](https://github.com/azavea/echo-locator/pull/694)
 - Add initial map styling and interactions [#690](https://github.com/azavea/echo-locator/pull/690)
+- Add formatted ranking lists to state [#698](https://github.com/azavea/echo-locator/pull/698)
 
 ### Changed